Extremely basic and non-functional Change Table, Call Service, Navbar
Still working on everything. Nothing does anything yet, but I have a implemented a nav bar that I still need to make functional and style correctly. I have also made a file for ChangeTable and Call for Service which also need to styled and functional. I also need to make it so that different components show up when they are supposed to. All I really have right now is text and useless buttons. Note that I have also done changes in the App.css file as a heads up.