Call Service and Change Table now function
Call service and change table will not interact with the user. Will need to update them as needed but the basic idea is there.
Call service allows the user to click a button to call service as well as cancel their request. It has 3 different states: call for the 1st time, cancel, call again. These states are looped through based on how many times the user clicks the button.
 Change table shows the different tables a user can select. Once selected by clicking the table is greyed out and a message tells the user which table they selected.
 A big issue now is the components will go back to their intialState when clicking on items in the navbar. My hope is that this will not be a huge issue if we make the initial state reflect the database. I will change as needed in the coming days/weeks.