ChangeTable saves state with LocalStorage
Now Change table will look to localStorage when determining its initialState. Also made it so that when you hover on the table the button changes color so the user is more aware that if they click that table that becomes the one they select. Still want to do more styling and other touches,